Add files via upload
Abstract
	
As house prices increase every year, so there is a need for implementing machine learning in a real estate to develop a model that predicts the current house prices. House prices prediction can help the developers to determine the estimated selling price of a house and can help the buyers to arrange the budget at the right time to purchase a house. There are many factors that influence the price of a house positively and some impact negatively such as the age of the house and include physical conditions, concept and location. This project presents a model that can generate predictions for housing prices by training the system with a 2001 data set so that it can learn how to forecast the market price of the house based on certain factors which have already mentioned. The connection between the houses costs and the economy of the nation is a critical factor for predicting house prices as housing prices trends are not only the concern of buyers and sellers, but it also plays major role in the current economic situation. Therefore, it is important to predict housing prices to help all the stack holders such as the buyers and the sellers to make their decisions accordingly. In order to select a prediction method, various regression methods were explored such as a multiple linear regression will be chosen for this research project due to its flexibility and probabilistic approach to learning and model prediction is very high.
Factors affecting the housing prices Including  house age, longitude, latitude, number of rooms, number of restrooms, number of schools around the house, carports, parks, shopping centers, condition of the house, total number of floors a house has and other distinctive accommodations around the house and besides a land territory are considered.

Methods used in this work are

1. Kmeans clustering algorithm
2. Multiple linear regression
     2.1. Step wise regression/backward elimination
3. Gradient Descent regression algorithm


 K-Means clustering 
The proposed work uses kmeans clustering algorithm to cluster the housing data into groups based on their feature similarities. In this work optimal number of clusters (K values) has been determined by using elbow curve method and after assigning k=3 Kmeans clustered the data into three clusters labelled as cluster 0, 1 and 2.

In figure 4.1. All the data in cluster 0 are coloured with red.
                       All the data in cluster 1 are coloured with blue.
                       All the data in cluster 2 are coloured with yellow.

Pseudo code for k-means clustering
Steps:   
             Input the numerical data
    1.  Decide the number of clusters k  in this work k=3 clusters
    2. Initially select the centroids at random k points, In sklearn machine learning library
         the KMeans++ algorithm is implemented in the background to select the random centroids    position
    3. Calculate the Euclidian distance between the data points and centroids
    4. Assign each data point to the nearest centroid to form a cluster d by means of minimum distance
    5. Reposition the centroids by calculating the mean value of all the points belonging to
             it, the mean value becomes the new value of the centroid.
    6. Reassign each data point to the closest centroid go to step 4 and repeat the process
             If there is no more reassignment go to step 7.
    7. Finally the model is in Convergence state and stop the process.
    8. Get the cluster column using from the result of the KMeans and append at the end of the housing data set that can be passed to Multiple linear regression

Pseudo code for Multiple Linear regression  
Step 1: Identify the dependent variable and independent variable (house price is dependent and all other variables are independent)
Step 2: Split the clustered data into train and set
Step 3: Train the model with train set 
Step 4 : Evaluate the goodness of model with unseen data called test set
Step 5. Measure how well the model is performing using R2 metrics
Step 6: Obtain the residual errors using MAE, SSE, RMSE metrics
Step 7: If the model is producing the desired result stop.
                Else
 Got to step 3 and use some other technique like gradient descent to improve the   performance or use regularization to overcome overfitting/underfitting problems which hinders the performance.



 Linear regression evaluation metrics used in the work are:-
    5. R Square (Coefficient of Determination) - R²= 1−SSE/SSTO
    6. Adjusted R²-
    7. MSE - MSE is the average of the squared error that is used as the loss function for linear regression.
    8. MAE - This is mean absolute error.

 Multiple Linear regression 
As MLR is a supervised learning algorithm the price is removed from the data set and assigned to y pandas data frame object, and all the dependent variables are assigned to x pandas data frame object. 
Splitting ratio used in this work is 80.20 which means 80% (1600)  of the observations are used for training the model and 20%(4001) of the observations are used for testing how well the model can predict the prices of the new houses which have not seen by the model in training phase


Deployemnt of the model 

To productionise the housing price prediction model so the clients can consume the prediction by providing their housing  data to the model via web page.
Bellow Html  page  is used to accept  user’s data and we have created Flask restful API to make the communication possible between the client and housing price prediction model.
Request from the client is handled  by the API and extracts the data from json object and pickled machine learning model in the server is unpickled and then the received data are passed as parameters to the model to make prediction and finally it returns a price of the house as a result which is serialised into json object and sent to the client via the API.